+++ packages/composites/vesper/src/adapter.ts
@@ -1,5 +1,6 @@
 import { Builder } from '@chainlink/ea-bootstrap'
-import { ExecuteWithConfig, ExecuteFactory, AdapterRequest, APIEndpoint } from '@chainlink/types'
+import * as TA from '@chainlink/token-allocation-adapter'
+import { ExecuteWithConfig, ExecuteFactory } from '@chainlink/types'
 import { makeConfig, Config } from './config'
 import * as endpoints from './endpoint'
 
@@ -7,8 +8,7 @@ export const execute: ExecuteWithConfig<Config> = async (request, context, confi
   return Builder.buildSelector(request, context, config, endpoints)
 }
 
-export const endpointSelector = (request: AdapterRequest): APIEndpoint<Config> =>
-  Builder.selectEndpoint(request, makeConfig(), endpoints)
+export const endpointSelector = TA.makeEndpointSelector(makeConfig, endpoints, 'price')
 
 export const makeExecute: ExecuteFactory<Config> = (config) => {
   return async (request, context) => execute(request, context, config || makeConfig())