+++ packages/composites/linear-finance/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @chainlink/linear-finance-adapter
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/token-allocation-adapter@1.3.0
+
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/token-allocation-adapter@1.2.3
+
 ## 2.0.7
 
 ### Patch Changes