+++ packages/composites/por-indexer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @chainlink/por-indexer-adapter
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/proof-of-reserves-adapter@1.4.18
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/proof-of-reserves-adapter@1.4.17
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.0.17
 
 ### Patch Changes