+++ packages/composites/set-token-index/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @chainlink/set-token-index-adapter
 
+## 1.0.50
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/token-allocation-adapter@1.3.0
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.0.49
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/token-allocation-adapter@1.2.3
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.0.48
 
 ### Patch Changes