+++ packages/sources/1forge/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @chainlink/1forge-adapter
 
+## 1.4.19
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.4.18
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.4.17
 
 ### Patch Changes