+++ packages/sources/chain-reserve-wallet/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @chainlink/chain-reserve-wallet-adapter
 
+## 2.2.16
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 2.2.15
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 2.2.14
 
 ### Patch Changes