+++ packages/sources/gramchain/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @chainlink/example-source-adapter
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.0.7
 
 ### Patch Changes