+++ packages/sources/sochain/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @chainlink/sochain-adapter
 
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-factories@1.0.34
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.2.19
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-factories@1.0.33
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.2.18
 
 ### Patch Changes