+++ packages/sources/tiingo/CHANGELOG.md
@@ -1,5 +1,26 @@
 ## v1.1.0
 
+## 1.8.0
+
+### Minor Changes
+
+- 65085c58d: fixed websocket support for iex endpoint
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.7.8
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.7.7
 
 ### Patch Changes